Built a pretty generic report generator
DataReport defines an abstract data report something that you can write
to, generate, and finalize. It also defines the basic model structure
of what a report can be generated from.

The output of a report is left open to interpretation, which is mostly
by AbstractReportGenerator and OutputBuilder exist.

The only implementation right now is PlaintextReportGenerator. This
generator produces a specifically formatted string representation of
a Report. This type of report is good for things like email.